+++ NEWS.md
@@ -1,6 +1,6 @@
 # rio 0.5.14
 
- * Changes to `gather_attrs()` for haven 2.0.0 release.
+ * Changes to `gather_attrs()` for haven 2.0.0 release. (#188)
  * Fixed a bug that generated a superfluous warning in `import()`.
  * Some style guide changes to code.